move usage from BetterPhpDocParser ast nodes to AttributeAwarePhpDoc
