+++ packages/PHPUnit/src/Rector/Class_/AddSeeTestAnnotationRector.php
@@ -9,7 +9,7 @@ use Nette\Utils\Strings;
 use PhpParser\Node;
 use PhpParser\Node\Stmt\Class_;
 use PHPStan\PhpDocParser\Ast\PhpDoc\PhpDocTagNode;
-use Rector\BetterPhpDocParser\Attributes\Ast\PhpDoc\AttributeAwareGenericTagValueNode;
+use Rector\AttributeAwarePhpDoc\Ast\PhpDoc\AttributeAwareGenericTagValueNode;
 use Rector\PHPUnit\Composer\ComposerAutoloadedDirectoryProvider;
 use Rector\Rector\AbstractRector;
 use Rector\RectorDefinition\CodeSample;
