+++ packages/PHPUnit/src/Rector/Class_/SelfContainerGetMethodCallFromTestToInjectPropertyRector.php
@@ -8,7 +8,7 @@ use PhpParser\Node;
 use PhpParser\Node\Stmt\Class_;
 use PhpParser\Node\Stmt\Property;
 use PHPStan\PhpDocParser\Ast\PhpDoc\GenericTagValueNode;
-use Rector\BetterPhpDocParser\Attributes\Ast\PhpDoc\AttributeAwarePhpDocTagNode;
+use Rector\AttributeAwarePhpDoc\Ast\PhpDoc\AttributeAwarePhpDocTagNode;
 use Rector\BetterPhpDocParser\PhpDocInfo\PhpDocInfo;
 use Rector\PhpParser\Node\Manipulator\ClassManipulator;
 use Rector\PHPUnit\Manipulator\OnContainerGetCallManipulator;
