+++ packages/TypeDeclaration/src/PhpDocParser/ParamPhpDocNodeFactory.php
@@ -6,8 +6,8 @@ namespace Rector\TypeDeclaration\PhpDocParser;
 
 use PhpParser\Node\Param;
 use PHPStan\Type\Type;
-use Rector\BetterPhpDocParser\Attributes\Ast\PhpDoc\AttributeAwareParamTagValueNode;
-use Rector\BetterPhpDocParser\Attributes\Ast\PhpDoc\AttributeAwarePhpDocTagNode;
+use Rector\AttributeAwarePhpDoc\Ast\PhpDoc\AttributeAwareParamTagValueNode;
+use Rector\AttributeAwarePhpDoc\Ast\PhpDoc\AttributeAwarePhpDocTagNode;
 use Rector\NodeTypeResolver\StaticTypeMapper;
 use Rector\PhpParser\Node\Resolver\NameResolver;
